# index/pilosa: speedup index iterator