Each next call to index iterator was searching for the bucket and creating a new transaction. Now the bucket and transaction is created only once and reused during all its life.

Previously:

```
index.Next() 5.445086ms
index.Next() 4.958599ms
index.Next() 4.597905ms
index.Next() 5.394169ms
index.Next() 5.161109ms
```

Now:

```
index.Next() 3.711µs
index.Next() 24.377µs
index.Next() 11.213µs
index.Next() 9.873µs
index.Next() 7.679µs
index.Next() 5.182µs
index.Next() 3.655µs
index.Next() 6.272µs
index.Next() 3.734µs
```

Some tests with queries:

* `create index path on files using pilosa (file_path);` (2620377 rows)

| rows returned | before | after |
| --- | --- | --- |
|160|1.34s|0.10s|
|1342|18s|6.34s|

* `create index author on commits using pilosa (commit_author_name);` (1933 rows)

| rows returned | before | after |
| --- | --- | --- |
|419|3.05s|0.51s|
|142|1.86s|0.84s|