@kuba-- 2018-07-03
Perfecto!
Do we rollback transaction in some cases?