@erizocosmico 2018-07-03
Actually, do we need transactions at all? If any error occurs the index is marked as corrupted and can't be used so it does not really matter which state the index is in