@jfontan 2018-07-03
I believe we need to do transactions always in boltdb. Also in read only transactions Rollback must be called.

https://godoc.org/github.com/boltdb/bolt#Tx.Rollback

I'll make the transaction read only and use Rollback always.