@jfontan 2018-07-03
Transactions are now read only and `Rollback` is used to close them.